healthcare-agents/backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.*
uvicorn[standard]>=0.34.0
python-dotenv==1.0.0

# Google ADK and A2A dependencies
google-adk==0.2.*
deprecated>=1.2.18
google-cloud-firestore==2.16.*
google-cloud-storage==2.18.*
google-cloud-pubsub==2.21.*
google-genai>=1.9.0

# Additional AI/ML dependencies
openai==1.12.*
anthropic==0.18.*
pydantic==2.5.*

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1